Partial implementation of point render service
This is a partially working version of the point render service.
The skeleton framework of the api is implemented to render the points
however the implemenation of plane fitting isnt.

There are errors with the point rendering function, particularly with
the projection into the camera. The points shoudl render based on their
distance, and should take up a larger field of view when the camer is
moved closer. Currently they do not.

The front end is not implemented outside of a template. The JS needs
to be written to enable movement of the camera in the xy plane (setting
a rotation and a distance with the camrea pointing at the origin).
Also some sort of interaction with loading data would be nice. Along
with some plane fitting.